static_cast<int> of Geos
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@3065 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Geos/textextruded.cpp
@@ -57,7 +57,7 @@ void textextruded :: setDepth(float prec)
 {
   m_fontDepth = prec;
   if(!m_font)return;
-  m_font->Depth((int)m_fontDepth);
+  m_font->Depth(m_fontDepth);
   setFontSize();
   setModified();
 }
